fix: use filepath as findup directory
VS Code uses the current install directory as the current directory,
which causes the prettier plugin to be unable to find the configs.
This caused the plugin to do nothing in VSCode on save.

I did try to specify the absolute path directly in Windows,
but that did not seem to be a work around. There may be a bug in the
future related to that in the future, but this should take
care of all reasonable use cases since an absolutely resolved
tsconfig/jsconfig.json would not be portable.
fix #1