fix: hack compatibility with other monkey-patched plugins
Hacky partial fix for #3.
This [prettier issue](https://github.com/prettier/prettier/issues/10261)
seems to be of the same concern.

No tests since this is complete garbage and hopefully temporary until
there is better support for this type of behavior, and since the current
fix seems to be considered an unintended usage of `options.plugins`.

The issue mentioned has other issues and a proposals linked,
after which it may make sense to revisit this issue.